// src/types/IProduct.ts
export interface IDownload {
  id: string;
  createdAt: string;
  updatedAt: string;
  sortOrder: number;
  title: string;
  summary: string;
  version: string;
  releasedOn: string;
  sizeBytes: string;
  downloadUrl: string;
  mirrors: { url: string; label: string }[] | null;
  platforms: string[];
  minOsVersion: string;
  fileType: string;
  sha256: string;
  deprecated: boolean;
  isActive: boolean;
  extra: {
    note?: string;
    language?: string[];
  };
}